add support for negotiating rowstride
note: the support is not yet complete, as some changes in the GstOmxBaseFilter will be required.  The basic issue (and also the reason that GstOmxBaseVideoDec needed to hard-code the color format in omx_setup()) is that the first buffers are not pad_alloc()'d.  So OMX component's output port is configured, and the component moved to executing, before the src caps are negotiated.  But changing this will be a significant enough change to merit a seperate commit.