Refactor buffer passing and allocation
Move all gst buffer/event <--> omx buffer translation into GOmxPort.  Including the buffer sharing logic and timestamp conversion.  Now for output ports, the gst element can set a buffer_alloc callback on the port object, which is used for to pad_alloc() replacement buffers (when buffer sharing is enabled), and also for allocating the initial buffers passed to OMX_UseBuffer() (again, only when buffer sharing is enabled).  This approach ensures that srcpad caps are negotiated before OMX_UseBuffer() (and therefore, before the OMX component is put in executing state).  This is required when the video sink element wants to use rowstride.